refactor: moved responses to nw mod
