+++ jarust/src/jaconnection.rs
@@ -5,10 +5,10 @@ use crate::nw::japrotocol::JaSuccessProtocol;
 use crate::nw::japrotocol::ResponseType;
 use crate::nw::jatransport::ConnectionParams;
 use crate::nw::jatransport::JaTransport;
+use crate::nw::respones::ServerInfoRsp;
 use crate::nw::transaction_gen::GenerateTransaction;
 use crate::params::CreateConnectionParams;
 use crate::prelude::*;
-use crate::respones::ServerInfoRsp;
 use jarust_transport::trans::TransportProtocol;
 use std::sync::Arc;
 use std::time::Duration;
