+++ jarust/src/lib.rs
@@ -7,7 +7,6 @@ pub mod jasession;
 pub mod nw;
 pub mod params;
 pub mod prelude;
-pub mod respones;
 
 pub use nw::transaction_gen::GenerateTransaction;
 pub use nw::transaction_gen::TransactionGenerationStrategy;
