+++ jarust/src/nw/japrotocol.rs
@@ -1,4 +1,4 @@
-use crate::respones::ServerInfoRsp;
+use super::respones::ServerInfoRsp;
 use serde::Deserialize;
 use serde::Serialize;
 use serde_json::Value;
